Better error handling of malformed filter parameters. Fixed such a bug in StripAttachmentsFilter.
git-svn-id: https://subetha.googlecode.com/svn/trunk@212 68df8246-1725-11de-bcda-2bb9902f8029

+++ plugin/src/org/subethamail/plugin/filter/StripAttachmentsFilter.java
@@ -44,7 +44,7 @@ public class StripAttachmentsFilter extends GenericFilter implements Lifecycle
 		new FilterParameterImpl(
 				"Threshold in K",
 				"Strip all attachments larger than this size, in kilobytes.  A value of 0 will strip all attachments.",
-				Long.class,
+				Integer.class,
 				100
 			)
 	};
